# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FastNEAR x402 facilitator for NEAR
  version: 0.1.0
  description: |
    API-key-gated x402 v2 exact-payment facilitator for Circle USDC on NEAR.
    Each deployed server is pinned to one network. Expected payment rejection
    is returned as an HTTP 200 protocol response; transport, authentication,
    policy, quota, conflict, and infrastructure errors use HTTP errors.
    X-API-Key is primary and Bearer is an alternative. If both are sent, they
    must contain the identical key; conflicting values receive HTTP 401.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: FastNEAR
    url: https://fastnear.com
servers:
  - url: https://test.x402.fastnear.com
    description: Testnet; not available until the launch checklist is complete
  - url: https://x402.fastnear.com
    description: Mainnet; not available until the launch checklist is complete
...
paths:
  /supported:
    get:
      tags: [Discovery]
      operationId: supported
      summary: Report the payment kind and extensions supported by this instance
      security: []
      responses:
        "200":
          description: Supported x402 capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportedResponse"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Primary resource-server credential; may accompany an identical Bearer value
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: x402 API key
      description: Alternative credential; may accompany an identical X-API-Key value
  responses:
...
    Unavailable:
      description: Database, leadership, reconciliation, RPC, or relayer is unavailable or indeterminate
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Suggested retry delay in seconds
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    NearNetwork:
      type: string
      enum:
        - near:testnet
        - near:mainnet
...
    SupportedKind:
      type: object
      required: [x402Version, scheme, network]
      additionalProperties: false
      properties:
        x402Version:
          type: integer
          const: 2
        scheme:
          type: string
          const: exact
        network:
          $ref: "#/components/schemas/NearNetwork"
        extra:
          type: object
          additionalProperties: true
    SupportedResponse:
      type: object
      required: [kinds, extensions, signers]
      additionalProperties: false
      properties:
        kinds:
          type: array
          minItems: 1
          maxItems: 1
          items:
            $ref: "#/components/schemas/SupportedKind"
        extensions:
          type: array
          contains:
            const: payment-identifier
          items:
            type: string
        signers:
          type: object
          description: Exact CAIP-2 network ID to facilitator signer account IDs
          additionalProperties:
            type: array
            items:
              type: string
...
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string